Move `samlang-pattern` into `samlang-nodes`
