+++ samlang-core-ast/samlang-nodes.ts
@@ -16,7 +16,6 @@ import {
   TypedComment,
 } from './common-nodes';
 import type { BinaryOperator } from './common-operators';
-import type { Pattern } from './samlang-pattern';
 
 interface BaseExpression extends Node {
   /** Identity of the object used for pattern matching. */
@@ -148,6 +147,40 @@ export interface LambdaExpression extends BaseExpression {
   readonly captured: Record<string, Type>;
   readonly body: SamlangExpression;
 }
+
+export interface TuplePattern extends Node {
+  readonly type: 'TuplePattern';
+  readonly destructedNames: readonly {
+    readonly name?: string;
+    readonly type: Type;
+    readonly range: Range;
+  }[];
+}
+
+export interface ObjectPatternDestucturedName {
+  readonly fieldName: string;
+  readonly fieldNameRange: Range;
+  readonly fieldOrder: number;
+  readonly type: Type;
+  readonly alias?: readonly [string, Range];
+  readonly range: Range;
+}
+
+export interface ObjectPattern extends Node {
+  readonly type: 'ObjectPattern';
+  readonly destructedNames: readonly ObjectPatternDestucturedName[];
+}
+
+export interface VariablePattern extends Node {
+  readonly type: 'VariablePattern';
+  readonly name: string;
+}
+
+export interface WildCardPattern extends Node {
+  readonly type: 'WildCardPattern';
+}
+
+export type Pattern = TuplePattern | ObjectPattern | VariablePattern | WildCardPattern;
 
 export interface SamlangValStatement extends Node {
   readonly pattern: Pattern;
